+++ src/generic/misc.h
@@ -16,7 +16,7 @@ void timer_kick(void);
 void *dynmem_start(void);
 void *dynmem_end(void);
 
-uint16_t crc16_ccitt(uint8_t *buf, uint_fast8_t len);
+uint16_t crc16_ccitt(uint8_t *buf, uint32_t len);
 
 void bootloader_request(void);
